This commit is a big code dump from a repository change.
Background: I switched to Fossil for a time -- I believed that it would make the repository better. However, fossil proved insufficient, compounded by the fact that there was nothing like GitHub for it. After a lor of
modifications to it trying to get the repository to work within the confines and differences of Fossil, I decided to move back to Git. This commit
packages all of those changes that I made during my time with Fossil.

Signed-off-by: Ethin Probst <ethindp@protonmail.com>